wrap stacktrace functions in StacktraceUtils
