+++ v2/cf-nodejs-logging-support/src/lib/logger/record-factory.ts
@@ -2,12 +2,12 @@ import util from "util";
 import Config from "../config/config";
 import ReqContext from "./context";
 import { SourceUtils } from "./source-utils";
+import { StacktraceUtils } from "./stacktrace-utils";
 const stringifySafe = require('json-stringify-safe');
 
 export default class RecordFactory {
 
     private static instance: RecordFactory;
-    private MAX_STACKTRACE_SIZE = 55 * 1024;
     private REDACTED_PLACEHOLDER = "redacted";
 
     private constructor() { }
@@ -34,11 +34,11 @@ export default class RecordFactory {
         var lastArg = args[args.length - 1];
 
         if (typeof lastArg === "object") {
-            if (this.isErrorWithStacktrace(lastArg)) {
-                record.stacktrace = this.prepareStacktrace(lastArg.stack);
+            if (StacktraceUtils.isErrorWithStacktrace(lastArg)) {
+                record.stacktrace = StacktraceUtils.prepareStacktrace(lastArg.stack);
             } else if (this.isValidObject(lastArg)) {
-                if (this.isErrorWithStacktrace(lastArg._error)) {
-                    record.stacktrace = this.prepareStacktrace(lastArg._error.stack);
+                if (StacktraceUtils.isErrorWithStacktrace(lastArg._error)) {
+                    record.stacktrace = StacktraceUtils.prepareStacktrace(lastArg._error.stack);
                     delete lastArg._error;
                 }
                 customFieldsFromArgs = lastArg;
@@ -114,55 +114,6 @@ export default class RecordFactory {
 
 
         return record;
-    }
-
-    // check if the given object is an Error with stacktrace using duck typing
-    private isErrorWithStacktrace(obj: any): boolean {
-        if (obj && obj.stack && obj.message && typeof obj.stack === "string" && typeof obj.message === "string") {
-            return true;
-        }
-        return false;
-    }
-
-    // Split stacktrace into string array and truncate lines if required by size limitation
-    // Truncation strategy: Take one line from the top and two lines from the bottom of the stacktrace until limit is reached.
-    private prepareStacktrace(stacktraceStr: any): any {
-        var fullStacktrace = stacktraceStr.split('\n');
-        var totalLineLength = fullStacktrace.reduce((acc: any, line: any) => acc + line.length, 0);
-
-        if (totalLineLength > this.MAX_STACKTRACE_SIZE) {
-            var truncatedStacktrace = [];
-            var stackA = [];
-            var stackB = [];
-            var indexA = 0;
-            var indexB = fullStacktrace.length - 1;
-            var currentLength = 73; // set to approx. character count for "truncated" and "omitted" labels
-
-            for (let i = 0; i < fullStacktrace.length; i++) {
-                if (i % 3 == 0) {
-                    let line = fullStacktrace[indexA++];
-                    if (currentLength + line.length > this.MAX_STACKTRACE_SIZE) {
-                        break;
-                    }
-                    currentLength += line.length;
-                    stackA.push(line);
-                } else {
-                    let line = fullStacktrace[indexB--];
-                    if (currentLength + line.length > this.MAX_STACKTRACE_SIZE) {
-                        break;
-                    }
-                    currentLength += line.length;
-                    stackB.push(line);
-                }
-            }
-
-            truncatedStacktrace.push("-------- STACK TRACE TRUNCATED --------");
-            truncatedStacktrace = [...truncatedStacktrace, ...stackA];
-            truncatedStacktrace.push(`-------- OMITTED ${fullStacktrace.length - (stackA.length + stackB.length)} LINES --------`);
-            truncatedStacktrace = [...truncatedStacktrace, ...stackB.reverse()];
-            return truncatedStacktrace;
-        }
-        return fullStacktrace;
     }
 
     private isValidObject(obj: any, canBeEmpty?: any): boolean {
